Decouple Typist from Player and Spawner
Typist no longer needs reference to player. Typist becomes a child of
player. So if there is no player, there is no typist.

Text Generation moved to Game singleton. For now, only a single Text
Generator is required for the entire game. Spawners can share the same
one.

Spawn signals are connected to typist so they are registered.